// server_api/src/modules/cycles/stages/stage_projects/stage.project.dto.ts
import mongoose from "mongoose";
import { ProjectStageStatus } from "./stage.project.enum";

export interface GetProjectStagesDTO {
    project?: mongoose.Types.ObjectId;
    stage?: mongoose.Types.ObjectId;
    status?: ProjectStageStatus;
    skip?: number;
    limit?: number;
}

export interface CreateProjectStageDTO {
    stage: mongoose.Types.ObjectId;
    project: mongoose.Types.ObjectId;
    documentPath: string;
    status?: ProjectStageStatus;
}

export interface UpdateProjectStageDTO {
    id: string;
    data: Partial<{
        documentPath: string;
        status: ProjectStageStatus;
    }>;
}

export interface DeleteProjectStageDto {
    id: string | mongoose.Types.ObjectId;
    userId: string;
}